+++ app/services/insights/average_time/for_merge_service.rb
@@ -30,11 +30,16 @@ module Insights
       end
 
       def calculate_merge_seconds(pull_request)
-        return pull_request.pull_merged_at.to_i - pull_request.pull_created_at.to_i unless @insightable.with_work_time?
+        created_at = pull_request.pull_created_at
+        merged_at = pull_request.pull_merged_at
+        # if PR merge was made before PR changed state from draft to open
+        # in such cases time spend for merge is 1 second
+        return 1 if created_at >= merged_at
+        return merged_at.to_i - created_at.to_i unless @insightable.with_work_time?
 
         seconds_between_times(
-          convert_time(pull_request.pull_created_at),
-          convert_time(pull_request.pull_merged_at)
+          convert_time(created_at),
+          convert_time(merged_at)
         )
       end
     end