+++ spec/services/import/savers/pull_requests_spec.rb
@@ -9,7 +9,7 @@ describe Import::Savers::PullRequests, type: :service do
     [
       {
         pull_number: 3,
-        pull_created_at: '2011-04-11T20:09:31Z',
+        pull_created_at: nil,
         pull_closed_at: nil,
         pull_merged_at: nil,
         author: {
@@ -58,6 +58,13 @@ describe Import::Savers::PullRequests, type: :service do
       expect { service_call }.to change(repository.pull_requests, :count).by(2)
     end
 
+    it 'one of the PRs has nil value for pull_created_at', :aggregate_failures do
+      service_call
+
+      expect(repository.pull_requests.find_by(pull_number: 3).pull_created_at).to be_nil
+      expect(repository.pull_requests.find_by(pull_number: 2).pull_created_at).not_to be_nil
+    end
+
     it 'creates 2 new entities' do
       expect { service_call }.to change(Entity, :count).by(2)
     end
@@ -68,6 +75,54 @@ describe Import::Savers::PullRequests, type: :service do
 
     it 'creates 2 reviewer PR entity' do
       expect { service_call }.to change(PullRequests::Entity.reviewer, :count).by(2)
+    end
+
+    context 'for receiving PRs with changing state from draft to open for review' do
+      let(:second_data) {
+        [
+          {
+            pull_number: 3,
+            pull_created_at: '2011-04-10T20:09:31Z',
+            pull_closed_at: nil,
+            pull_merged_at: nil,
+            author: {
+              external_id: 1,
+              provider: Providerable::GITHUB,
+              login: 'octocat',
+              avatar_url: 'https://github.com/images/error/octocat_happy.gif',
+              html_url: 'https://github.com/octocat'
+            },
+            reviewers: [
+              {
+                external_id: 1,
+                provider: Providerable::GITHUB,
+                login: 'octocat',
+                avatar_url: 'https://github.com/images/error/octocat_happy.gif',
+                html_url: 'https://github.com/octocat'
+              },
+              {
+                external_id: 2,
+                provider: Providerable::GITHUB,
+                login: 'octocat2',
+                avatar_url: 'https://github.com/images/error/octocat_happy.gif',
+                html_url: 'https://github.com/octocat'
+              }
+            ]
+          }
+        ]
+      }
+
+      it 'sets pull_created_at to current time', :aggregate_failures do
+        service_call
+
+        pull_request = repository.pull_requests.find_by(pull_number: 3)
+        expect(pull_request.pull_created_at).to be_nil
+
+        described_class.call(repository: repository, data: second_data)
+
+        expect(pull_request.reload.pull_created_at).not_to be_nil
+        expect(pull_request.pull_created_at.to_date).not_to eq DateTime.new(2011, 4, 10, 20, 9, 31)
+      end
     end
 
     context 'when there is existing entity' do