+++ spec/services/insights/average_time/for_merge_service_spec.rb
@@ -31,6 +31,30 @@ describe Insights::AverageTime::ForMergeService, type: :service do
     it 'succeeds' do
       expect(service_call.success?).to be_truthy
     end
+
+    context 'when some PR does not have pull_created_at' do
+      before { pr2.update!(pull_created_at: nil) }
+
+      it 'generates average time without such PRs' do
+        expect(service_call.result).to eq({})
+      end
+
+      it 'succeeds' do
+        expect(service_call.success?).to be_truthy
+      end
+    end
+
+    context 'when some PR has pull_created_at after reviewed' do
+      before { pr2.update!(pull_created_at: DateTime.new(2023, 1, 4, 1, 0, 0)) }
+
+      it 'generates less average time for such PRs' do
+        expect(service_call.result).to eq({ entity2.id => 1 })
+      end
+
+      it 'succeeds' do
+        expect(service_call.success?).to be_truthy
+      end
+    end
   end
 
   context 'for company insightable' do